update jai compiler version
I cannot build mooviz for the web anymore, which is quite a problem. It seems to be having more corruption issues like what the WASM crash fixer was supposed to fix, but that isn't doing it anymore... But I don't have the time to worry about it at the moment, so for now we will just have to stick to the old build.